+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+### v1.0.0 - *5/6/2014*
+- Composer support
+- PSR-4 compliant
+- Refactored/renamed using camelCase
+- Fixed examples. Freegeoip.net renamed response property `areacode` to `area_code`.
+
 ### v0.1.3 - *3/9/2014*
 - Fixed serious bug in **Curl.php**. Was not setting curl option `CURLOPT_CUSTOMREQUEST` back to `GET` and `POST` in `get_request()` and `post_request()`.